Break up massive utils file
