+++ python_modules/dagster/dagster/_core/definitions/freshness_checks/time_partition.py
@@ -8,10 +8,10 @@ from dagster._core.definitions.time_window_partitions import TimeWindowPartition
 from ..asset_checks import AssetChecksDefinition
 from ..assets import AssetsDefinition, SourceAsset
 from ..events import CoercibleToAssetKey
+from .shared_builder import build_freshness_checks_for_assets
 from .utils import (
     DEFAULT_FRESHNESS_SEVERITY,
     DEFAULT_FRESHNESS_TIMEZONE,
-    build_freshness_checks_for_assets,
 )
